tests: Added skip messages to tests
Went through all the tests in layer_validation_tests and added skip
messages to all tests that conditionally skip their checks.
Additionally, added the prefix "TEST_SKIPPED" to all the skip messages
so that they can easily be parsed.

Also added a script that looks parses the output of all the tests and
prints a short summary of which tests were skipped on every device
profile and which tests print unexpected output.